Move the ObjectSubstitution class from deployment to runtime module issue #573
